06_GUI/requirements.txt
flask
mysql-connector-python
authlib
requests
python-dotenv
